Allow customization of additional files
Support for more advanced files (e.g. docs and config) is provided
via the "file_extras" plugin.

For simple file copies and backwards compatibility, colon-delimited
values are still supported.

Additionally, a `dict` format is supported. The required keys are
`src` and `dest` which behave the same as in the legacy syntax. To
mark a file as a configuration file, the `config` key should be `True`.
The same applies for documentation files (with the key `doc`).

If the value for the `config` key is a string, this will be used to
further customize the directive. This is typically used with
`noreplace` or `missingok`, which will produce `%config(noreplace)` and
`%config(missingok)` respectively.

Example configuration (JSON):
```
{
  "file_extras": {
    "files": [
      "bar:opt/foo/bar",
      {
        "src": "readme.txt",
        "dest": "usr/share/doc/foobar/readme.txt",
        "doc": true
      },
      {
        "src": "app.conf",
        "dest": "etc/foobar/app.conf",
        "config": "noreplace"
      }
    ]
  }
}
```