Update venvctrl for nu shell support
This removes the workaround in the tests that was used to skip nu shell
installation in the virtualenv and updates tests to use the latest
venvctrl version.

The latest version of venvctrl also delcares a dependency on virtualenv
so it no longer needs to be manually installed in the CI configuration.