Update test dependencies and remove pylint
Pylint has not run in the automated test suite for a while. It is now
removed from the local test suite that runs using tox.

The test-requirements are updated to resolve more bit-rot related
dependency issues. In this case, one of the pinned test dependencies
required an older version of a transitive dependency that failed to run
on the latest versions of Python.

The test suite now runs using the active Python interpreter version
rather than being pinned to 3.9. Prior to this, all tests would attempt
to use a local python3.9 installation for the virtualenv created during
the tests. The new test code always uses the current active interpreter
so that test environments do not have to support multiple Python
versions.