Remove nushell warnings in README
There's no change to rpmvenv but a recent patch to venvctrl has removed
the need for nushell related workarounds. Users pulling in the 0.6.0 or
later version of venvctrl will automatically have the patch applied.
Users no longer need to disable nushell for rpmvenv to work. This may
change again in the future if virtualenv makes more changes to
activation scripts.